ingredient-parser-service/requirements.txt
ingredient-parser-nlp==2.1.1
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6